+++ src/templates/vite-react/src/components/ui/LinkButton.tsx
@@ -1,44 +1,11 @@
-import type { LinkProps } from 'react-router-dom';
+import { Link, type LinkProps } from 'react-router-dom';
 import type { VariantProps } from 'class-variance-authority';
 import { buttonVariants } from './Button';
-import { Link, type Params, type Path } from '@/router';
 import { cn } from '@/utils/styles';
 
-type LinkButtonProps<
-  P extends
-    | Path
-    | {
-        pathname: Path;
-        search?: string | undefined;
-        hash?: string | undefined;
-      },
-> = LinkProps &
-  VariantProps<typeof buttonVariants> &
-  (P extends keyof Params
-    ? {
-        to: P;
-        params: Params[P];
-      }
-    : P extends {
-          pathname: Path;
-          search?: string | undefined;
-          hash?: string | undefined;
-        }
-      ? P['pathname'] extends keyof Params
-        ? {
-            to: P;
-            params: Params[P['pathname']];
-          }
-        : {
-            to: P;
-            params?: undefined;
-          }
-      : {
-          to: P;
-          params?: undefined;
-        });
+type LinkButtonProps = LinkProps & VariantProps<typeof buttonVariants>;
 
-const LinkButton = <P extends Path>(props: LinkButtonProps<P>) => (
+const LinkButton = (props: LinkButtonProps) => (
   <Link
     {...props}
     className={cn(
