+++ src/templates/vite-react/src/main.tsx
@@ -1,7 +1,6 @@
-import { startTransition, StrictMode, Suspense } from 'react';
+import { lazy, startTransition, StrictMode, Suspense } from 'react';
 import { createRoot } from 'react-dom/client';
 import { ErrorBoundary } from 'react-error-boundary';
-import { Routes } from '@generouted/react-router/lazy';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import env from './env';
@@ -10,6 +9,8 @@ import './styles/index.css';
 import './i18n';
 import PageError from './components/app/PageError';
 import LoadingOverlay from './components/ui/LoadingOverlay';
+
+const Routes = lazy(() => import('@/routes/Routes'));
 
 const prepare = async () => {
   if (env.MODE === 'development' && env.VITE_APP_MSW_ENABLED) {
