+++ src/templates/vite-react/src/routes/404/route.tsx
@@ -1,0 +1,23 @@
+import { Button } from '@/components/ui/Button';
+import PageWrapper from '@/components/app/PageWrapper';
+import useAppTranslation from '@/i18n/useAppTranslation';
+import { useNavigate } from 'react-router-dom';
+
+const NotFoundPage = () => {
+  const { t } = useAppTranslation();
+  const navigate = useNavigate();
+
+  const handleOnClick = () => {
+    navigate('/');
+  };
+
+  return (
+    <PageWrapper>
+      <h1>{t('NotFoundPage.notFound')}</h1>
+      <p>{t('NotFoundPage.pleaseTryADifferentRoute')}</p>
+      <Button onClick={handleOnClick}>{t('NotFoundPage.home')}</Button>
+    </PageWrapper>
+  );
+};
+
+export default NotFoundPage;
