+++ src/templates/vite-react/src/routes/react-query/[id]/route.tsx
@@ -1,10 +1,10 @@
 import PageWrapper from '@/components/app/PageWrapper';
-import { useParams } from '@/router';
 import useGetPostQuery from '@/services/queries/useGetPostQuery';
+import { useParams } from 'react-router-dom';
 
 const ReactQueryPostPage = () => {
-  const { id } = useParams('/react-query/:id');
-  const { data, isLoading, isError } = useGetPostQuery(id);
+  const { id } = useParams();
+  const { data, isLoading, isError } = useGetPostQuery(id!);
 
   return (
     <PageWrapper isLoading={isLoading} isError={isError}>
