+++ src/templates/vite-react/vite.config.ts
@@ -6,11 +6,10 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import svgr from 'vite-plugin-svgr';
 import checker from 'vite-plugin-checker';
-import generouted from '@generouted/react-router/plugin';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), svgr(), checker({ typescript: true }), generouted()],
+  plugins: [react(), svgr(), checker({ typescript: true })],
   resolve: {
     alias: {
       '@': path.resolve(__dirname, './src'),
